move bus vhdl generating code to bus_vhdl_gen.py
